+++ pkg/api/handlers/channel_type_handler_test.go
@@ -33,7 +33,6 @@ type channelTypeAPITest struct {
 func channelTypeDICases(funcName string) []channelTypeAPITest {
 	parsedCTDI, _ := json.Marshal(models.ChannelTypeDI{
 		ChannelType: meta.ChannelType{Meta: meta.Metadata{Name: "mock_channelType"}},
-		Scope:       "",
 		DryRun:      false,
 	})
 	wrongFormatData := []byte{1}
@@ -113,7 +112,6 @@ func channelTypeDICases(funcName string) []channelTypeAPITest {
 // use these test cases
 func channelTypeQueryDICases(funcName string) []channelTypeAPITest {
 	parsedCTQDI, _ := json.Marshal(models.ChannelTypeQueryDI{
-		Scope:  "",
 		CtName: "mock_channelType",
 		DryRun: false,
 	})