+++ pkg/cmd/utils/cli_client.go
@@ -6,7 +6,6 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/inspr/inspr/pkg/auth"
 	"github.com/inspr/inspr/pkg/cmd"
 	"github.com/inspr/inspr/pkg/controller"
 	"github.com/inspr/inspr/pkg/controller/client"
@@ -58,13 +57,14 @@ func SetClient(url string) {
 		dir, _ := os.UserHomeDir()
 		cmd.InsprOptions.Token = filepath.Join(dir, ".inspr/token")
 	}
+
 	config := client.ControllerConfig{
-		Auth: auth.Authenticator{
-			TokenPath: cmd.InsprOptions.Token,
+		Auth: Authenticator{
+			cmd.InsprOptions.Token,
 		},
-		Scope: "",
-		URL:   url,
+		URL: url,
 	}
+
 	defaults.client = client.NewControllerClient(config)
 }