+++ pkg/controller/client/channel.go
@@ -5,34 +5,32 @@ import (
 
 	"github.com/inspr/inspr/pkg/api/models"
 	"github.com/inspr/inspr/pkg/meta"
-	metautils "github.com/inspr/inspr/pkg/meta/utils"
 	"github.com/inspr/inspr/pkg/meta/utils/diff"
+	"github.com/inspr/inspr/pkg/rest"
 	"github.com/inspr/inspr/pkg/rest/request"
 )
 
 // ChannelClient interacts with channels on the Insprd
 type ChannelClient struct {
-	client *request.Client
-	config ControllerConfig
+	reqClient *request.Client
 }
 
 // Get gets a channel from the Insprd
 //
-// The context refers to the parent app of the given channel, represented with a dot separated query
+// The scope refers to the parent app of the given channel, represented with a dot separated query
 // such as app1.app2
 //
 // The name is the name of the channel. So to search for a channel inside app1 with the name channel1 you
 // would call cc.Get(context.Background(), "app1", "channel1")
-func (cc *ChannelClient) Get(ctx context.Context, context string, name string) (*meta.Channel, error) {
-	fullscope, _ := metautils.JoinScopes(cc.config.Scope, context)
+func (cc *ChannelClient) Get(ctx context.Context, scope string, name string) (*meta.Channel, error) {
 	cdi := models.ChannelQueryDI{
-		Scope:  fullscope,
 		ChName: name,
 	}
-
 	var resp meta.Channel
 
-	err := cc.client.Send(ctx, "/channels", "GET", cdi, &resp)
+	err := cc.reqClient.
+		Header(rest.HeaderScopeKey, scope).
+		Send(ctx, "/channels", "GET", cdi, &resp)
 	if err != nil {
 		return nil, err
 	}
@@ -42,23 +40,23 @@ func (cc *ChannelClient) Get(ctx context.Context, context string, name string) (
 
 // Create creates a channel inside the Insprd
 //
-// The context refers to the parent app of the given channel, represented with a dot separated query
+// The scope refers to the parent app of the given channel, represented with a dot separated query
 // such as **app1.app2**
 //
 // The channel information such as name and etc will be inferred from the given channel's metadata.
 //
 // So to create a channel inside app1 with the name channel1 you
 // would call cc.Create(context.Background(), "app1", &meta.Channel{...})
-func (cc *ChannelClient) Create(ctx context.Context, context string, ch *meta.Channel, dryRun bool) (diff.Changelog, error) {
-	fullscope, _ := metautils.JoinScopes(cc.config.Scope, context)
+func (cc *ChannelClient) Create(ctx context.Context, scope string, ch *meta.Channel, dryRun bool) (diff.Changelog, error) {
 	cdi := models.ChannelDI{
-		Scope:   fullscope,
 		Channel: *ch,
 		DryRun:  dryRun,
 	}
-
 	var resp diff.Changelog
-	err := cc.client.Send(ctx, "/channels", "POST", cdi, &resp)
+
+	err := cc.reqClient.
+		Header(rest.HeaderScopeKey, scope).
+		Send(ctx, "/channels", "POST", cdi, &resp)
 	if err != nil {
 		return nil, err
 	}
@@ -68,23 +66,23 @@ func (cc *ChannelClient) Create(ctx context.Context, context string, ch *meta.Ch
 
 // Delete deletes a channel inside the Insprd
 //
-// The context refers to the parent app of the given channel, represented with a dot separated query
+// The scope refers to the parent app of the given channel, represented with a dot separated query
 // such as **app1.app2**
 //
 // The name is the name of the channel to be deleted.
 //
 // So to delete a channel inside app1 with the name channel1 you
 // would call cc.Delete(context.Background(), "app1", "channel1")
-func (cc *ChannelClient) Delete(ctx context.Context, context string, name string, dryRun bool) (diff.Changelog, error) {
-	fullscope, _ := metautils.JoinScopes(cc.config.Scope, context)
+func (cc *ChannelClient) Delete(ctx context.Context, scope string, name string, dryRun bool) (diff.Changelog, error) {
 	cdi := models.ChannelQueryDI{
-		Scope:  fullscope,
 		ChName: name,
 		DryRun: dryRun,
 	}
-
 	var resp diff.Changelog
-	err := cc.client.Send(ctx, "/channels", "DELETE", cdi, &resp)
+
+	err := cc.reqClient.
+		Header(rest.HeaderScopeKey, scope).
+		Send(ctx, "/channels", "DELETE", cdi, &resp)
 	if err != nil {
 		return nil, err
 	}
@@ -94,23 +92,23 @@ func (cc *ChannelClient) Delete(ctx context.Context, context string, name string
 
 // Update creates a channel inside the Insprd
 //
-// The context refers to the parent app of the given channel, represented with a dot separated query
+// The scope refers to the parent app of the given channel, represented with a dot separated query
 // such as **app1.app2**
 //
 // The channel information such as name and etc will be inferred from the given channel's metadata.
 //
 // So to update a channel inside app1 with the name channel1 you
 // would call cc.Update(context.Background(), "app1", &meta.Channel{...})
-func (cc *ChannelClient) Update(ctx context.Context, context string, ch *meta.Channel, dryRun bool) (diff.Changelog, error) {
-	fullscope, _ := metautils.JoinScopes(cc.config.Scope, context)
+func (cc *ChannelClient) Update(ctx context.Context, scope string, ch *meta.Channel, dryRun bool) (diff.Changelog, error) {
 	cdi := models.ChannelDI{
-		Scope:   fullscope,
 		Channel: *ch,
 		DryRun:  dryRun,
 	}
-
 	var resp diff.Changelog
-	err := cc.client.Send(ctx, "/channels", "PUT", cdi, &resp)
+
+	err := cc.reqClient.
+		Header(rest.HeaderScopeKey, scope).
+		Send(ctx, "/channels", "PUT", cdi, &resp)
 	if err != nil {
 		return nil, err
 	}