+++ pkg/controller/client/channel_test.go
@@ -12,15 +12,16 @@ import (
 	"github.com/inspr/inspr/pkg/ierrors"
 	"github.com/inspr/inspr/pkg/meta"
 	"github.com/inspr/inspr/pkg/meta/utils/diff"
+	"github.com/inspr/inspr/pkg/rest"
 	"github.com/inspr/inspr/pkg/rest/request"
 )
 
 func TestChannelClient_Delete(t *testing.T) {
 
 	type args struct {
-		ctx     context.Context
-		context string
-		name    string
+		ctx   context.Context
+		scope string
+		name  string
 	}
 	tests := []struct {
 		name    string
@@ -30,16 +31,16 @@ func TestChannelClient_Delete(t *testing.T) {
 		{
 			name: "delete channel test",
 			args: args{
-				ctx:     context.Background(),
-				context: "app1.app2",
+				ctx:   context.Background(),
+				scope: "app1.app2",
 			},
 			wantErr: false,
 		},
 		{
 			name: "delete channel with error test",
 			args: args{
-				ctx:     context.Background(),
-				context: "app1.app2",
+				ctx:   context.Background(),
+				scope: "app1.app2",
 			},
 			wantErr: true,
 		},
@@ -63,6 +64,7 @@ func TestChannelClient_Delete(t *testing.T) {
 				}
 
 				var di models.ChannelQueryDI
+				scope := r.Header.Get(rest.HeaderScopeKey)
 
 				decoder := request.JSONDecoderGenerator(r.Body)
 				err := decoder.Decode(&di)
@@ -70,8 +72,12 @@ func TestChannelClient_Delete(t *testing.T) {
 					t.Error(err)
 				}
 
-				if di.Scope != tt.args.context {
-					t.Errorf("context set incorrectly. want = %v, got = %v", di.Scope, tt.args.context)
+				if scope != tt.args.scope {
+					t.Errorf(
+						"context set incorrectly. want = %v, got = %v",
+						scope,
+						tt.args.scope,
+					)
 				}
 				if di.ChName != tt.args.name {
 					t.Errorf("name set incorrectly. want = %v, got = %v", di.ChName, tt.args.name)
@@ -82,9 +88,9 @@ func TestChannelClient_Delete(t *testing.T) {
 			s := httptest.NewServer(http.HandlerFunc(handler))
 			defer s.Close()
 			ac := &ChannelClient{
-				client: request.NewJSONClient(s.URL),
+				reqClient: request.NewJSONClient(s.URL),
 			}
-			if _, err := ac.Delete(tt.args.ctx, tt.args.context, tt.args.name, false); (err != nil) != tt.wantErr {
+			if _, err := ac.Delete(tt.args.ctx, tt.args.scope, tt.args.name, false); (err != nil) != tt.wantErr {
 				t.Errorf("ChannelClient.Delete() error = %v, wantErr %v", err, tt.wantErr)
 			}
 		})
@@ -149,6 +155,7 @@ func TestChannelClient_Get(t *testing.T) {
 				}
 
 				var di models.ChannelQueryDI
+				scope := r.Header.Get(rest.HeaderScopeKey)
 
 				decoder := request.JSONDecoderGenerator(r.Body)
 				err := decoder.Decode(&di)
@@ -156,11 +163,19 @@ func TestChannelClient_Get(t *testing.T) {
 					t.Error(err)
 				}
 
-				if di.Scope != tt.args.context {
-					t.Errorf("context set incorrectly. want = %v, got = %v", di.Scope, tt.args.context)
+				if scope != tt.args.context {
+					t.Errorf(
+						"context set incorrectly. want = %v, got = %v",
+						scope,
+						tt.args.context,
+					)
 				}
 				if di.ChName != tt.args.name {
-					t.Errorf("name set incorrectly. want = %v, got = %v", di.ChName, tt.args.name)
+					t.Errorf(
+						"name set incorrectly. want = %v, got = %v",
+						di.ChName,
+						tt.args.name,
+					)
 				}
 
 				encoder.Encode(tt.want)
@@ -169,7 +184,7 @@ func TestChannelClient_Get(t *testing.T) {
 			s := httptest.NewServer(http.HandlerFunc(handler))
 			defer s.Close()
 			ac := &ChannelClient{
-				client: request.NewJSONClient(s.URL),
+				reqClient: request.NewJSONClient(s.URL),
 			}
 			got, err := ac.Get(tt.args.ctx, tt.args.context, tt.args.name)
 			if (err != nil) != tt.wantErr {
@@ -239,6 +254,7 @@ func TestChannelClient_Create(t *testing.T) {
 				}
 
 				var di models.ChannelDI
+				scope := r.Header.Get(rest.HeaderScopeKey)
 
 				decoder := request.JSONDecoderGenerator(r.Body)
 				err := decoder.Decode(&di)
@@ -246,8 +262,12 @@ func TestChannelClient_Create(t *testing.T) {
 					t.Error(err)
 				}
 
-				if di.Scope != tt.args.context {
-					t.Errorf("context set incorrectly. want = %v, got = %v", di.Scope, tt.args.context)
+				if scope != tt.args.context {
+					t.Errorf(
+						"context set incorrectly. want = %v, got = %v",
+						scope,
+						tt.args.context,
+					)
 				}
 
 				if !reflect.DeepEqual(di.Channel, *tt.args.ch) {
@@ -258,7 +278,7 @@ func TestChannelClient_Create(t *testing.T) {
 			s := httptest.NewServer(http.HandlerFunc(handler))
 			defer s.Close()
 			ac := &ChannelClient{
-				client: request.NewJSONClient(s.URL),
+				reqClient: request.NewJSONClient(s.URL),
 			}
 			if _, err := ac.Create(tt.args.ctx, tt.args.context, tt.args.ch, false); (err != nil) != tt.wantErr {
 				t.Errorf("ChannelClient.Create() error = %v, wantErr %v", err, tt.wantErr)
@@ -324,6 +344,7 @@ func TestChannelClient_Update(t *testing.T) {
 				}
 
 				var di models.ChannelDI
+				scope := r.Header.Get(rest.HeaderScopeKey)
 
 				decoder := request.JSONDecoderGenerator(r.Body)
 				err := decoder.Decode(&di)
@@ -331,8 +352,12 @@ func TestChannelClient_Update(t *testing.T) {
 					t.Error(err)
 				}
 
-				if di.Scope != tt.args.context {
-					t.Errorf("context set incorrectly. want = %v, got = %v", di.Scope, tt.args.context)
+				if scope != tt.args.context {
+					t.Errorf(
+						"context set incorrectly. want = %v, got = %v",
+						scope,
+						tt.args.context,
+					)
 				}
 
 				if !reflect.DeepEqual(di.Channel, *tt.args.ch) {
@@ -343,7 +368,7 @@ func TestChannelClient_Update(t *testing.T) {
 			s := httptest.NewServer(http.HandlerFunc(handler))
 			defer s.Close()
 			ac := &ChannelClient{
-				client: request.NewJSONClient(s.URL),
+				reqClient: request.NewJSONClient(s.URL),
 			}
 			if _, err := ac.Update(tt.args.ctx, tt.args.context, tt.args.ch, false); (err != nil) != tt.wantErr {
 				t.Errorf("ChannelClient.Update() error = %v, wantErr %v", err, tt.wantErr)