+++ pkg/rest/middleware.go
@@ -1,16 +1,63 @@
 package rest
 
 import (
-	"bytes"
-	"encoding/json"
-	"io/ioutil"
 	"log"
 	"net/http"
 	"strings"
 
 	"github.com/inspr/inspr/pkg/auth"
 	"github.com/inspr/inspr/pkg/ierrors"
+	"github.com/inspr/inspr/pkg/utils"
 )
+
+// CRUDHandler handles crud requests to a given resource
+type CRUDHandler interface {
+	HandleCreate() Handler
+	HandleDelete() Handler
+	HandleUpdate() Handler
+	HandleGet() Handler
+	GetAuth() auth.Auth
+}
+
+// HandleCRUD uses a CRUDHandler to handle HTTP requests for a CRUD resource
+func HandleCRUD(handler CRUDHandler) Handler {
+	return func(w http.ResponseWriter, r *http.Request) {
+		switch r.Method {
+
+		case http.MethodGet:
+			handler.
+				HandleGet().
+				Validate(handler.GetAuth()).
+				JSON().
+				Recover()(w, r)
+
+		case http.MethodPost:
+			handler.
+				HandleCreate().
+				Validate(handler.GetAuth()).
+				JSON().
+				Recover()(w, r)
+
+		case http.MethodPut:
+			handler.
+				HandleUpdate().
+				Validate(handler.GetAuth()).
+				JSON().
+				Recover()(w, r)
+
+		case http.MethodDelete:
+			handler.
+				HandleDelete().
+				Validate(handler.GetAuth()).
+				JSON().
+				Recover()(w, r)
+
+		default:
+			http.Error(w, "405 method not allowed", http.StatusMethodNotAllowed)
+			return
+		}
+	}
+}
 
 // JSON specifies in the header that the response content is a json
 func (h Handler) JSON() Handler {
@@ -56,72 +103,65 @@ func (h Handler) Validate(auth auth.Auth) Handler {
 			return
 		}
 
-		// request scope
-		requestData := struct {
-			Scope string `json:"scope"`
-		}{}
-
-		// Read the content
-		if r.Body != nil {
-			// reads body
-			bodyBytes, _ := ioutil.ReadAll(r.Body)
-			// unmarshal into scope Data
-			json.Unmarshal(bodyBytes, &requestData)
-			// Restore the r.Body to its original state
-			r.Body = ioutil.NopCloser(bytes.NewBuffer(bodyBytes))
-		}
-
+		// used for checking scope authorization
+		reqScopes := r.Header[HeaderScopeKey]
 		log.Printf("payload.Permissions = %+v\n", payload.Permissions)
-		valid := false
+
+		// used for checking permissions
+		operation := getOperation(r)
+		target := getTarget(r)
+		perm := operation + ":" + target
+
 		for scope := range payload.Permissions {
-			log.Printf("scope = %+v\n", scope)
-			if strings.HasPrefix(requestData.Scope, scope) {
-				// scope found
-				valid = true
+			log.Printf("permission-scope = %+v\n", scope)
+
+			// usually the request will one have one scope
+			for _, rs := range reqScopes {
+				log.Printf("request-scope = %+v\n", rs)
+
+				if strings.HasPrefix(rs, scope) &&
+					utils.Includes(payload.Permissions[scope], perm) {
+					// token and context are valid
+					h(w, r)
+					return
+				}
 			}
 		}
 
-		// check for unauthorized error
-		if !valid {
-			ERROR(w, ierrors.NewError().Forbidden().Message("not enought permissions to perform request").Build())
-			return
-
-		}
-
-		// token and context are valid
-		h(w, r)
+		// there were no valid operations
+		ERROR(
+			w,
+			ierrors.
+				NewError().
+				Forbidden().
+				Message("not enought permissions to perform request").
+				Build(),
+		)
 	}
 }
 
-// CRUDHandler handles crud requests to a given resource
-type CRUDHandler interface {
-	HandleCreate() Handler
-	HandleDelete() Handler
-	HandleUpdate() Handler
-	HandleGet() Handler
-	GetAuth() auth.Auth
+// getOperation returns the operation being done by the Request in the cluster
+// get, create, update, delete.
+func getOperation(r *http.Request) string {
+	// some methods represent their own operation
+	operation, ok := operationTranslator[r.Method]
+	if !ok {
+		return r.Method
+	}
+	return operation
 }
 
-// HandleCRUD uses a CRUDHandler to handle HTTP requests for a CRUD resource
-func HandleCRUD(handler CRUDHandler) Handler {
-	return func(w http.ResponseWriter, r *http.Request) {
-		switch r.Method {
+// getTarget isolates the area that is being requested, for example the request
+// URL is https://example.org:8000/channels, the getTarget removes the base of the url and some unecessary '/' and returns only 'channels'
+func getTarget(r *http.Request) string {
+	route := strings.TrimSuffix(r.URL.Path, "/")
+	route = strings.TrimPrefix(route, r.URL.Host)
+	route = strings.TrimPrefix(route, "/")
 
-		case http.MethodGet:
-			handler.HandleGet().Validate(handler.GetAuth()).JSON().Recover()(w, r)
-
-		case http.MethodPost:
-			handler.HandleCreate().Validate(handler.GetAuth()).JSON().Recover()(w, r)
-
-		case http.MethodPut:
-			handler.HandleUpdate().Validate(handler.GetAuth()).JSON().Recover()(w, r)
-
-		case http.MethodDelete:
-			handler.HandleDelete().Validate(handler.GetAuth()).JSON().Recover()(w, r)
-
-		default:
-			http.Error(w, "405 method not allowed", http.StatusMethodNotAllowed)
-			return
-		}
+	// some constant values differ from the url name used
+	target, ok := routeTranslator[route]
+	if !ok {
+		return route
 	}
+	return target
 }