+++ pkg/rest/middleware_test.go
@@ -2,12 +2,14 @@ package rest
 
 import (
 	"bytes"
-	"encoding/json"
 	"errors"
+	"fmt"
 	"io/ioutil"
+	"log"
 	"net/http"
 	"net/http/httptest"
 	"reflect"
+	"strings"
 	"testing"
 
 	"github.com/inspr/inspr/pkg/auth"
@@ -62,10 +64,13 @@ func TestHandler_JSON(t *testing.T) {
 }
 
 func TestHandler_Validate(t *testing.T) {
+	log.SetOutput(ioutil.Discard)
 	type args struct {
-		auth        auth.Auth
-		headerValue string
-		scope       string
+		auth          auth.Auth
+		Authorization string
+		httpMethod    string
+		reqURLSuffix  string
+		scope         string
 	}
 	tests := []struct {
 		name     string
@@ -76,7 +81,8 @@ func TestHandler_Validate(t *testing.T) {
 		{
 			name: "no_auth_header",
 			args: args{
-				auth: authMock.NewMockAuth(nil),
+				auth:       authMock.NewMockAuth(nil),
+				httpMethod: http.MethodPost,
 			},
 			wantErr:  true,
 			wantCode: http.StatusUnauthorized,
@@ -87,7 +93,8 @@ func TestHandler_Validate(t *testing.T) {
 				auth: authMock.NewMockAuth(
 					&ierrors.InsprError{Code: ierrors.InvalidToken},
 				),
-				headerValue: "Bearer mock_token",
+				Authorization: "Bearer mock_token",
+				httpMethod:    http.MethodPost,
 			},
 			wantErr:  true,
 			wantCode: http.StatusUnauthorized,
@@ -98,7 +105,8 @@ func TestHandler_Validate(t *testing.T) {
 				auth: authMock.NewMockAuth(
 					errors.New("mock_error"),
 				),
-				headerValue: "Bearer mock_token",
+				Authorization: "Bearer mock_token",
+				httpMethod:    http.MethodPost,
 			},
 			wantErr:  true,
 			wantCode: http.StatusInternalServerError,
@@ -106,9 +114,10 @@ func TestHandler_Validate(t *testing.T) {
 		{
 			name: "scope_error",
 			args: args{
-				auth:        authMock.NewMockAuth(nil),
-				headerValue: "Bearer mock_token",
-				scope:       "no_valid_scope",
+				auth:          authMock.NewMockAuth(nil),
+				Authorization: "Bearer mock_token",
+				scope:         "no_valid_scope",
+				httpMethod:    http.MethodPost,
 			},
 			wantErr:  true,
 			wantCode: http.StatusForbidden,
@@ -116,9 +125,22 @@ func TestHandler_Validate(t *testing.T) {
 		{
 			name: "scope_error_doesnt_have_prefix",
 			args: args{
-				auth:        authMock.NewMockAuth(nil),
-				headerValue: "Bearer mock_token",
-				scope:       "wrongScope.scope_1",
+				auth:          authMock.NewMockAuth(nil),
+				Authorization: "Bearer mock_token",
+				scope:         "wrongScope.scope_1",
+				httpMethod:    http.MethodPost,
+			},
+			wantErr:  true,
+			wantCode: http.StatusForbidden,
+		},
+		{
+			name: "no_permission_for_operation",
+			args: args{
+				auth:          authMock.NewMockAuth(nil),
+				Authorization: "Bearer mock_token",
+				scope:         "scope_1",
+				httpMethod:    http.MethodGet,
+				reqURLSuffix:  "channels",
 			},
 			wantErr:  true,
 			wantCode: http.StatusForbidden,
@@ -126,9 +148,11 @@ func TestHandler_Validate(t *testing.T) {
 		{
 			name: "working",
 			args: args{
-				auth:        authMock.NewMockAuth(nil),
-				headerValue: "Bearer mock_token",
-				scope:       "scope_1",
+				auth:          authMock.NewMockAuth(nil),
+				Authorization: "Bearer mock_token",
+				scope:         "scope_1",
+				httpMethod:    http.MethodPost,
+				reqURLSuffix:  "channels",
 			},
 			wantErr:  false,
 			wantCode: http.StatusOK,
@@ -141,22 +165,18 @@ func TestHandler_Validate(t *testing.T) {
 			defer ts.Close()
 			client := ts.Client()
 
-			req, err := http.NewRequest(http.MethodPost, ts.URL, nil)
+			reqURL := fmt.Sprintf("%s/%s", ts.URL, tt.args.reqURLSuffix)
+			reqURL = strings.TrimSuffix(reqURL, "/")
+
+			req, err := http.NewRequest(tt.args.httpMethod, reqURL, nil)
 			if err != nil {
 				t.Error("error creating request")
 			}
 
 			// adds auth to request header
-			req.Header.Add("Authorization", tt.args.headerValue)
-
-			// scope to body
-			scopeData := struct {
-				Scope string `json:"scope"`
-			}{
-				Scope: tt.args.scope,
-			}
-			scopeBytes, _ := json.Marshal(scopeData)
-			req.Body = ioutil.NopCloser(bytes.NewBuffer(scopeBytes))
+			req.Header.Add("Authorization", tt.args.Authorization)
+			// adds the scope to request header
+			req.Header.Add("Scope", tt.args.scope)
 
 			// does request
 			res, err := client.Do(req)
@@ -172,6 +192,118 @@ func TestHandler_Validate(t *testing.T) {
 					got,
 					tt.wantCode,
 				)
+			}
+		})
+	}
+}
+
+func Test_getOperation(t *testing.T) {
+	type args struct {
+		r *http.Request
+	}
+	createReq := func(method string) *http.Request {
+		req, _ := http.NewRequest(
+			method,
+			"mock-url",
+			bytes.NewBuffer([]byte{}),
+		)
+		return req
+	}
+
+	tests := []struct {
+		name string
+		args args
+		want string
+	}{
+		{
+			name: "GetOperation",
+			args: args{
+				r: createReq(http.MethodGet),
+			},
+			want: "get",
+		},
+		{
+			name: "CreateOperation",
+			args: args{
+				r: createReq(http.MethodPost),
+			},
+			want: "create",
+		},
+		{
+			name: "UpdateOperation",
+			args: args{
+				r: createReq(http.MethodPut),
+			},
+			want: "update",
+		},
+		{
+			name: "DeleteOperation",
+			args: args{
+				r: createReq(http.MethodDelete),
+			},
+			want: "delete",
+		},
+		{
+			name: "InvalidOperation",
+			args: args{
+				r: createReq(http.MethodPatch),
+			},
+			want: http.MethodPatch,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := getOperation(tt.args.r); got != tt.want {
+				t.Errorf("getOperation() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func Test_getTarget(t *testing.T) {
+	type args struct {
+		r *http.Request
+	}
+	const baseURL = "https://localhost:8080/"
+	createReq := func(url string) *http.Request {
+		req, _ := http.NewRequest(
+			http.MethodPost,
+			baseURL+url,
+			bytes.NewBuffer([]byte{}),
+		)
+		return req
+	}
+
+	tests := []struct {
+		name string
+		args args
+		want string
+	}{
+		{
+			name: "simple_target",
+			args: args{createReq("hello")},
+			want: "hello",
+		},
+		{
+			name: "composed_target",
+			args: args{createReq("hello/world/")},
+			want: "hello/world",
+		},
+		{
+			name: "exception_dapps",
+			args: args{createReq("apps")},
+			want: "dapp",
+		},
+		{
+			name: "exception_channeltypes",
+			args: args{createReq("channeltypes")},
+			want: "ctype",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := getTarget(tt.args.r); got != tt.want {
+				t.Errorf("getTarget() = %v, want %v", got, tt.want)
 			}
 		})
 	}